Refactoring changes: move enums to the base file, have specific implementations in separate files
